+++ lib/src/screens/initial_setup_widgets/res_create_form_submission.dart
@@ -33,7 +33,7 @@ library;
 import 'package:flutter/material.dart';
 
 import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:solidpod/solidpod.dart' show initPod;
+import 'package:solidpod/solidpod.dart' show KeyManager;
 
 import 'package:solidui/src/constants/initial_setup.dart';
 import 'package:solidui/src/widgets/solid_animation_dialog.dart';
@@ -167,14 +167,13 @@ ElevatedButton resCreateFormSubmission(
         final securityKey = formData[securityKeyStr].toString();
 
         try {
-          // await _initPodOriginalFunc(securityKey);
-          await initPod(
-            securityKey,
-            dirUrls: resFoldersLink,
-            fileUrls: resFilesLink,
-          );
+          // Initialise the POD with the security key.
+          // Set up the KeyManager and create verification keys.
+
+          await KeyManager.initPodKeys(securityKey);
+          debugPrint('POD initialised with security key');
         } on Exception catch (e) {
-          debugPrint(e.toString());
+          debugPrint('Error initialising POD: $e');
         }
 
         await Navigator.pushReplacement(