+++ CHANGELOG.md
@@ -17,7 +17,7 @@ au](https://solidcommunity.au/docs/solidui)
 
 + Remove solidui/solidpod circular dependency [0.0.7 20251029 gjw]
 + Refactor SOlidLogin for max 300 loc lint [0.0.6 20251017 cdawei]
-+ Fix security key handling [0.0.5 20251027 tchecn]
++ Fix security key handling [0.0.5 20251027 tchen]
 + Add webid to an info header in navdrawer [0.0.4 20251017 tchen]
 + Update version widget dependency [0.0.3 20251008 gjw]
 + EXAMPLE: Add demo of getting version [0.0.2 20250820 gjw]